requirements.txt
click==6.6
Flask==0.11.1
requests==2.25.1
PyJWT==2.1.0
jwcrypto==0.8
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
marshmallow==2.10.3
nose2==0.6.5
six==1.16.0
Werkzeug==0.11.11
flask-dynamo==0.1.2
flask-cors==3.0.10
google-analytics-data==0.5.0
oauth2client==4.1.3
google-api-python-client==2.4.0
facebook_business==10.0.1
boto3==1.17.39
simplejson=3.17.2